+++ morecantile/models.py
@@ -1389,15 +1389,14 @@ class TileMatrixSet(BaseModel, arbitrary_types_allowed=True):
                 UserWarning,
                 stacklevel=1,
             )
+
             if authority_code := feature_crs.to_authority(min_confidence=20):
                 authority, code = authority_code
                 feat.update(
                     {
                         "crs": {
                             "type": "name",
-                            "properties": {
-                                "name": f"urn:ogc:def:crs:{authority}:0:{code}"
-                            },
+                            "properties": {"name": CRS_to_uri(feature_crs)},
                         }
                     }
                 )